infra/app/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
pandas==2.2.3
boto3==1.36.21
botocore==1.36.21
s3transfer==0.11.2
mlflow==2.20.2
pyspark==3.5.4
findspark==2.0.1
python-dotenv==1.0.1
